sign up login

+++ app/Http/Controllers/TasksController.php
@@ -14,9 +14,15 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if(\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = ['user' => $user,'tasks' => $tasks,];
+        }
         
-        return view('tasks.index', ['tasks' => $tasks, ]);
+        return view('tasks.index', $data);
     }
 
     /**
@@ -40,10 +46,8 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, ['status' => 'required | max:10',  'content' => 'required | max:191', ]);
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        
+        $request->user()->tasks()->create(['content' => $request->content, 'status' => $request ->status,]);
         
         return redirect('/');
     }
@@ -57,8 +61,9 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-        
-        return view('tasks.show', ['tasks.show', 'task' => $task, ]);
+        if (\Auth::id() === $task->user_id) {
+        return view('tasks.show', ['tasks.show', 'task' => $task, ]); }
+        return redirect('/');
     }
 
     /**
@@ -84,9 +89,10 @@ class TasksController extends Controller
     {
         $this->validate($request, ['status' => 'required | max:10',  'content' => 'required | max:191', ]);
         $task = Task::find($id);
+        if (\Auth::id() === $task->user_id) {
         $task->content = $request->content;
         $task->status = $request->status;
-        $task->save();
+        $task->save();}
         return redirect('/');
     }
 
@@ -99,7 +105,8 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {$task->delete();}
         
         return redirect('/');
     }
